+++ i21Database/dbo/Tables/tblIPEntityStage.sql
@@ -22,7 +22,7 @@
 	[strImportStatus] NVARCHAR(50) COLLATE Latin1_General_CI_AS NULL,  
 	[strSessionId] NVARCHAR(50) COLLATE Latin1_General_CI_AS NULL,
 	dtmTransactionDate DATETIME NULL  DEFAULT((getdate())),
-	intTrxSequenceNo INT,
+	intTrxSequenceNo BIGINT,
 	strCompanyLocation NVARCHAR(6) COLLATE Latin1_General_CI_AS,
 	intActionId INT,
 	strStatus NVARCHAR(50) COLLATE Latin1_General_CI_AS,
