+++ i21Database/dbo/Tables/tblIPItemRouteDetailArchive.sql
@@ -6,8 +6,8 @@
 	strManufacturingCell NVARCHAR(50) COLLATE Latin1_General_CI_AS,
 	strStorageLocation NVARCHAR(50) COLLATE Latin1_General_CI_AS,
 	strManufacturingGroup NVARCHAR(50) COLLATE Latin1_General_CI_AS,
-	intTrxSequenceNo INT,
-	intParentTrxSequenceNo INT,
+	intTrxSequenceNo BIGINT,
+	intParentTrxSequenceNo BIGINT,
 
 	CONSTRAINT [PK_tblIPItemRouteDetailArchive] PRIMARY KEY (intItemRouteDetailStageId),
 	CONSTRAINT [FK_tblIPItemRouteDetailArchive_tblIPItemRouteArchive] FOREIGN KEY (intItemRouteStageId) REFERENCES tblIPItemRouteArchive(intItemRouteStageId) ON DELETE CASCADE
